# swagger.yaml
openapi: 3.0.0
info:
  title: OpenRouteService API
  description: API endpoints to calculate travel time and routes using OpenRouteService
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
paths:
  /travel-time:
    get:
      summary: Get travel time between two points
      parameters:
        - name: origin
          in: query
          required: true
          schema:
            type: string
            example: '78.0081,27.1767'
          description: Origin coordinates in format 'longitude,latitude'
        - name: destination
          in: query
          required: true
          schema:
            type: string
            example: '77.2090,28.6139'
          description: Destination coordinates in format 'longitude,latitude'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  travelTimeHours:
                    type: number
                    format: float
                    description: Travel time in hours
                    example: 2.5
                  travelDistanceKM:
                    type: number
                    format: float
                    description: Travel distance in kilometers
                    example: 200.5
        '400':
          description: Invalid coordinates format or missing parameters
        '500':
          description: Error calculating travel time
  /route:
    get:
      summary: Calculate route with multiple destinations
      parameters:
        - name: locations
          in: query
          required: true
          schema:
            type: string
            example: '27.1767,78.0081|28.6139,77.2090|...'
          description: Pipe-separated list of coordinates (latitude,longitude)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  travelTimeSeconds:
                    type: integer
                    description: Total travel time in seconds
                    example: 7200
                  travelDistanceMeters:
                    type: integer
                    description: Total travel distance in meters
                    example: 150000
        '400':
          description: Invalid coordinates format or missing parameters
        '500':
          description: Error calculating route